Move ping plugin into folder
