+++ src/headless/headless.js
@@ -13,7 +13,7 @@ import "./plugins/disco/index.js";  // XEP-0030 Service discovery
 import "./plugins/headlines.js";    // Support for headline messages
 import "./plugins/mam/index.js";    // XEP-0313 Message Archive Management
 import "./plugins/muc/index.js";    // XEP-0045 Multi-user chat
-import "./plugins/ping.js";         // XEP-0199 XMPP Ping
+import "./plugins/ping/index.js";         // XEP-0199 XMPP Ping
 import "./plugins/pubsub.js";       // XEP-0060 Pubsub
 import "./plugins/roster/index.js"; // RFC-6121 Contacts Roster
 import "./plugins/smacks.js";       // XEP-0198 Stream Management
